requirements.txt
fastapi==0.103.1
uvicorn==0.23.1
pydantic==2.3.0
sqlalchemy==2.0.20
databases==0.9.0
scikit-learn==1.3.0
python-dotenv==1.0.0
